test: split the dataset builders out of conftest into tests/fixtures
conftest.py had grown to 875 lines holding 35 plain factory functions
alongside its 20 fixtures, which made it hard to answer the question a
reader of a test actually has: where did this data come from, and what
does it contain.

The factories move into a `tests/fixtures/` package grouped by what they
build, and test modules import from the specific module rather than
through a re-export, so the origin is visible at the point of use.
The fixtures stay in conftest.py, which pytest still discovers as before.

No `pytest_plugins` registration is involved: these are plain functions,
so ordinary imports are enough. Every moved function is byte-identical to
its previous definition, and collected test counts are unchanged both in
total and per module.
