+++ tests/fixtures/__init__.py
@@ -1,0 +1,5 @@
+"""Dataset builders shared across the test suite.
+
+Import directly from the module that owns a builder, so a reader of a
+test can see where its data comes from.
+"""
